+++ lib/enums/device_screen_type.dart
