+++ lib/responsive/sizing_information.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:responsive_architecture/enums/device_screen_type.dart';
+
+class SizingInformation {
+  final DeviceScreenType deviceScreenType;
+  final Size screenSize;
+  final Size localWidgetSize;
+
+  SizingInformation({
+    this.deviceScreenType,
+    this.screenSize,
+    this.localWidgetSize,
+  });
+
+  @override
+  String toString() {
+    return 'DeviceType:$deviceScreenType ScreenSize:$screenSize LocalWidgetSize:$localWidgetSize';
+  }
+}
