--- lib/ui/sizing_information.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/widgets.dart';
-import 'package:responsive_architecture/enums/device_screen_type.dart';
-
-class SizingInformation {
-  final Orientation orientation;
-  final DeviceScreenType deviceType;
-  final Size screenSize;
-  final Size localWidgetSize;
-
-  SizingInformation({
-    this.orientation,
-    this.deviceType,
-    this.screenSize,
-    this.localWidgetSize,
-  });
-
-  @override
-  String toString() {
-    return 'Orientation:$orientation DeviceType:$deviceType ScreenSize:$screenSize LocalWidgetSize:$localWidgetSize';
-  }
-}
