--- lib/ui/views/home_view.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:responsive_architecture/ui/base_widget.dart';
-
-class HomeView extends StatelessWidget {
-  const HomeView({Key key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BaseWidget(
-      builder: (context, sizingInformation) {
-        return Scaffold(
-          body: Center(
-            child: Column(
-              children: [
-                Container(
-                  height: 150,
-                  margin: EdgeInsets.all(50),
-                  color: Colors.blue,
-                  child: BaseWidget(
-                    builder: (context, sizingInformation) =>
-                        Text(sizingInformation.toString()),
-                  ),
-                ),
-                Text(sizingInformation.toString()),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
