+++ lib/utils/ui_utils.dart
@@ -2,15 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:responsive_architecture/enums/device_screen_type.dart';
 
 DeviceScreenType getDeviceType(MediaQueryData mediaQuery) {
-  var orientation = mediaQuery.orientation;
-
-  double deviceWidth = 0;
-
-  if (orientation == Orientation.landscape) {
-    deviceWidth = mediaQuery.size.height;
-  } else {
-    deviceWidth = mediaQuery.size.width;
-  }
+  double deviceWidth = mediaQuery.size.shortestSide;
 
   if (deviceWidth > 950) {
     return DeviceScreenType.Desktop;
