+++ lib/views/home/home_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomeView extends StatelessWidget {
+  HomeView({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+    );
+  }
+}
